requirements.txt
asgiref==3.11.0
chardet==5.2.0
colorama==0.4.6
decorator==5.2.1
Django==6.0.1
django-cors-headers==4.9.0
djangorestframework==3.16.1
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.4.1
pillow==11.3.0
proglog==0.1.12
pydub==0.25.1
pysrt==1.1.2
python-dotenv==1.2.1
SpeechRecognition==3.14.5
sqlparse==0.5.5
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
